Remove out-of-date description in Auto3DSeg FAQ and fix anchors (#1282)
The Auto3DSeg FAQ environment setup description looks out-of-date. It is
a duplicate because each notebook has the installation "Setup
environment" section. To avoid maintaining these texts in the tutorial
for the future, I recommend a deletion as shown in this PR.

Also, some anchors won't work properly in Jupyter Lab 2.x. The fixes are
also included.

### Description
A few sentences describing the changes proposed in this pull request.

### Checks
<!--- Put an `x` in all the boxes that apply, and remove the not
applicable items -->
- [x] Avoid including large-size files in the PR.
- [x] Clean up long text outputs from code cells in the notebook.
- [x] For security purposes, please check the contents and remove any
sensitive info such as user names and private key.
- [x] Ensure (1) hyperlinks and markdown anchors are working (2) use
relative paths for tutorial repo files (3) put figure and graphs in the
`./figure` folder

---------

Signed-off-by: Mingxin Zheng <18563433+mingxin-zheng@users.noreply.github.com>